+++ integration-tests/src/screens/PropertiesScreen.tsx
@@ -28,6 +28,24 @@ export const PropertiesScreen = ({navigation}) => {
         title="Clear Event Properties"
         onPress={() => Heap.clearEventProperties()}
       />
+      <Button
+        title="Log getUserId"
+        onPress={async () =>
+          Heap.track('getUserId', {value: await Heap.getUserId()})
+        }
+      />
+      <Button
+        title="Log getSessionId"
+        onPress={async () =>
+          Heap.track('getSessionId', {value: await Heap.getSessionId()})
+        }
+      />
+      <Button
+        title="Send Custom Event"
+        onPress={async () =>
+          Heap.track('custom-event', { a: {b: null, c: [1,2,3], d: new (class Test {}) } })
+        }
+      />
       <Button title="Back" onPress={() => navigation.goBack()} />
     </ScrollView>
   );