+++ integration-tests/tests/util/types.ts
@@ -6,6 +6,10 @@ export interface CaptureSourceEvent extends CaptureEvent {
   sourceEvent: CaptureSourceEventPayload;
 }
 
+export interface CaptureSourceCustomEvent extends CaptureEvent {
+  sourceCustomEvent: CaptureSourceCustomEventPayload;
+}
+
 export interface BoolProperty {
   bool: boolean;
 }
@@ -19,6 +23,11 @@ export interface CaptureSourceEventPayload {
   name?: string;
   sourceProperties: {[key: string]: BoolProperty | StringProperty};
   source: string;
+}
+
+export interface CaptureSourceCustomEventPayload
+  extends CaptureSourceEventPayload {
+  customProperties: {[key: string]: BoolProperty | StringProperty};
 }
 
 export interface CaptureMessage<T extends CaptureEvent> {
@@ -48,9 +57,15 @@ export interface CaptureAndroidUserMigration {
 }
 
 export function isSourceEvent(
-  event: CaptureEvent,
+  event: CaptureEvent | null,
 ): event is CaptureSourceEvent {
   return has(event, 'sourceEvent');
+}
+
+export function isSourceCustomEvent(
+  event: CaptureEvent | null,
+): event is CaptureSourceCustomEvent {
+  return has(event, 'sourceCustomEvent');
 }
 
 // This keeps getting formatted away. :(
@@ -78,7 +93,7 @@ export function getPropertyValue(
 
 export function matcherForSourceEventWithProperties(
   expectedType: string,
-  expectedProperties: {[key: string]: string | boolean},
+  expectedProperties: {[key: string]: string | boolean | RegExp},
 ): (
   message: CaptureMessage<CaptureEvent>,
 ) => message is CaptureMessage<CaptureSourceEvent> {
@@ -97,11 +112,76 @@ export function matcherForSourceEventWithProperties(
     }
 
     for (const [key, value] of Object.entries(expectedProperties)) {
-      if (getPropertyValue(key, event.sourceEvent.sourceProperties) !== value) {
+      if (
+        !validatePropertyValue(key, event.sourceEvent.sourceProperties, value)
+      ) {
         return false;
       }
     }
 
     return true;
   };
+}
+
+export function matcherForSourceCustomEventWithProperties(
+  expectedType: string,
+  expectedSourceProperties: {[key: string]: string | boolean | RegExp},
+  expectedCustomProperties: {[key: string]: string | boolean | RegExp},
+): (
+  message: CaptureMessage<CaptureEvent>,
+) => message is CaptureMessage<CaptureSourceCustomEvent> {
+  return (message): message is CaptureMessage<CaptureSourceCustomEvent> => {
+    const event = message.event;
+
+    if (!isSourceCustomEvent(event)) {
+      return false;
+    }
+
+    if (
+      event.sourceCustomEvent.type !== expectedType &&
+      event.sourceCustomEvent.name !== expectedType
+    ) {
+      return false;
+    }
+
+    for (const [key, value] of Object.entries(expectedSourceProperties)) {
+      if (
+        !validatePropertyValue(
+          key,
+          event.sourceCustomEvent.sourceProperties,
+          value,
+        )
+      ) {
+        return false;
+      }
+    }
+
+    for (const [key, value] of Object.entries(expectedCustomProperties)) {
+      if (
+        !validatePropertyValue(
+          key,
+          event.sourceCustomEvent.customProperties,
+          value,
+        )
+      ) {
+        return false;
+      }
+    }
+
+    return true;
+  };
+}
+
+function validatePropertyValue(
+  propertyName: string,
+  properties: {[key: string]: BoolProperty | StringProperty},
+  value: string | boolean | RegExp,
+): boolean {
+  const actualValue = getPropertyValue(propertyName, properties);
+
+  if (value instanceof RegExp) {
+    return typeof actualValue === 'string' && value.test(actualValue);
+  } else {
+    return actualValue === value;
+  }
 }